+++ packages/client/src/components/renote-button.vue
